Rfishmorph: FISHMORPH morphology, landmarks and quality control

Tools dedicated to the FISHMORPH morphological database (Brosse et al., 2021). The package reconstructs the 21 FISHMORPH landmarks, computes the 11 body segments and the 9 dimensionless ratios, standardizes landmark geometry and enforces the FISHMORPH digitizing conventions, projects specimens into the FISHMORPH functional trait space, and provides visualization and quality-control tools. It also offers a guided workflow to add new species, using rfishbase to validate taxonomy and build a global freshwater fish list.

Reference

Brosse, S., Charpin, N., Su, G., Toussaint, A., Herrera-R, G.A., Tedesco, P.A. & Villeger, S. (2021). FISHMORPH: A global database on morphological traits of freshwater fishes. Global Ecology and Biogeography, 30, 2330-2336. doi:10.1111/geb.13395

See also

Author

Maintainer: Aurele Toussaint aurele.toussaint@cnrs.fr (affiliation: CNRS)

Authors:

Aurele Toussaint aurele.toussaint@cnrs.fr (affiliation: CNRS)